+++ ez/ez.go
@@ -8,12 +8,12 @@ import (
 
 	"github.com/vimeo/dials"
 	"github.com/vimeo/dials/common"
+	"github.com/vimeo/dials/decoders/json"
+	"github.com/vimeo/dials/decoders/toml"
+	"github.com/vimeo/dials/decoders/yaml"
 	"github.com/vimeo/dials/sources/env"
 	"github.com/vimeo/dials/sources/file"
 	"github.com/vimeo/dials/sources/flag"
-	"github.com/vimeo/dials/sources/json"
-	"github.com/vimeo/dials/sources/toml"
-	"github.com/vimeo/dials/sources/yaml"
 	"github.com/vimeo/dials/sourcewrap"
 	"github.com/vimeo/dials/tagformat"
 	"github.com/vimeo/dials/tagformat/caseconversion"
